Make deleting recovery.conf optional.
https://github.com/zalando/patroni/issues/637

pgBackRest's restore command generates the appropriate recovery.conf based
on the parameters you provide to pgBackRest.  When calling pgBackRest's restore command
via Patroni's custom bootstrap, it deletes that recovery.conf.  Specifying the recovery.conf
information in the patroni.yml is less than ideal.  It prevent's leveraging pgBackRests
work to ensure recovery.conf files are properly generated.  It also can lead to transient
config data in the patroni.yml under certain restore cases, such as a PITR restore
of Cluster B to  Cluster A, where the restore_commnand in A needs to reference B.

The parameter is optional.  The default behavior is to delete the recovery.conf.

+++ patroni/postgresql.py
@@ -545,6 +545,10 @@ class Postgresql(object):
         self.set_state('running custom bootstrap script')
         params = ['--scope=' + self.scope, '--datadir=' + self._data_dir]
         try:
+            keep_existing_recovery_conf = config['keep_existing_recovery_conf']
+        except KeyError:
+            keep_existing_recovery_conf = False
+        try:
             logger.info('Running custom bootstrap script: %s', config['command'])
             if self.cancellable_subprocess_call(shlex.split(config['command']) + params) != 0:
                 self.set_state('custom bootstrap failed')
@@ -556,7 +560,7 @@ class Postgresql(object):
 
         if 'recovery_conf' in config:
             self.write_recovery_conf(config['recovery_conf'])
-        elif os.path.isfile(self._recovery_conf) or os.path.islink(self._recovery_conf):
+        elif (os.path.isfile(self._recovery_conf) or os.path.islink(self._recovery_conf)) and not keep_existing_recovery_conf:
             os.unlink(self._recovery_conf)
         return True
